[minor] Extract snapshot read
